+++ core/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @commercetools-test-data/core
 
+## 10.4.0
+
+## 10.3.0
+
 ## 10.2.0
 
 ## 10.1.4