+++ models/cart-discount/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @commercetools-test-data/cart-discount
 
+## 10.4.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @commercetools-test-data/core@10.4.0
+  - @commercetools-test-data/category@10.4.0
+  - @commercetools-test-data/commons@10.4.0
+  - @commercetools-test-data/customer-group@10.4.0
+  - @commercetools-test-data/product-type@10.4.0
+  - @commercetools-test-data/utils@10.4.0
+
+## 10.3.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @commercetools-test-data/core@10.3.0
+  - @commercetools-test-data/category@10.3.0
+  - @commercetools-test-data/commons@10.3.0
+  - @commercetools-test-data/customer-group@10.3.0
+  - @commercetools-test-data/product-type@10.3.0
+  - @commercetools-test-data/utils@10.3.0
+
 ## 10.2.0
 
 ### Patch Changes